app/lavish_backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.0.0
requests==2.31.0
django-environ==0.12.0
whitenoise==6.9.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.7
django-jazzmin==3.0.1
django-redis==6.0.0
django-timezone-field==7.1
psycopg2-binary==2.9.10
Pillow==11.3.0
celery==5.5.3
redis==6.2.0
gunicorn==23.0.0
